breaking(gatsby-source-contentful): remove gatsby-image support (#33528)
