+++ e2e-tests/contentful/snapshots.js
@@ -28,7 +28,7 @@ module.exports = {
   },
   "rich-text": {
     "rich-text: All Features": {
-      "1": "<div data-cy-id=\"rich-text-all-features\">\n  <h2>Rich Text: All Features</h2>\n  <h1>The <a href=\"https://contentful.com/\">European</a> languages</h1>\n  <p>are members of the same family. Their <u>separate</u> <strong\n      data-cy-strong=\"true\">existence</strong> is a <i>myth</i>. For:</p>\n  <ul>\n    <li>\n      <p>science</p>\n    </li>\n    <li>\n      <p>music</p>\n    </li>\n    <li>\n      <p>sport</p>\n    </li>\n    <li>\n      <p>etc</p>\n    </li>\n  </ul>\n  <p>Europe uses the same vocabulary.</p>\n  <div class=\" gatsby-image-wrapper\"\n    style=\"position:relative;overflow:hidden;width:200px\">\n    <div aria-hidden=\"true\"\n      style=\"width:100%;padding-bottom:150%\"></div><img aria-hidden=\"true\"\n      src=\"data:image/redacted;base64,redacted\"\n      alt=\"\"\n      style=\"position: absolute; top: 0px; left: 0px; width: 100%; height: 100%; object-fit: cover; object-position: center center; opacity: 0; transition-delay: 500ms;\">\n    <picture>\n      <source\n        srcset=\"https://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=50&amp;h=75&amp;q=50 50w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=100&amp;h=150&amp;q=50 100w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=200&amp;h=300&amp;q=50 200w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=300&amp;h=450&amp;q=50 300w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=400&amp;h=600&amp;q=50 400w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=600&amp;h=900&amp;q=50 600w\"\n        sizes=\"(max-width: 200px) 100vw, 200px\"><img\n        sizes=\"(max-width: 200px) 100vw, 200px\"\n        srcset=\"https://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=50&amp;h=75&amp;q=50 50w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=100&amp;h=150&amp;q=50 100w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=200&amp;h=300&amp;q=50 200w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=300&amp;h=450&amp;q=50 300w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=400&amp;h=600&amp;q=50 400w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=600&amp;h=900&amp;q=50 600w\"\n        src=\"https://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=200&amp;q=50\"\n        alt=\"\"\n        loading=\"lazy\"\n        style=\"position: absolute; top: 0px; left: 0px; width: 100%; height: 100%; object-fit: cover; object-position: center center; opacity: 1; transition: opacity 500ms ease 0s;\">\n    </picture><noscript>\n      <picture>\n        <source\n          srcset=\"https://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=50&h=75&q=50 50w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=100&h=150&q=50 100w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=200&h=300&q=50 200w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=300&h=450&q=50 300w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=400&h=600&q=50 400w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=600&h=900&q=50 600w\"\n          sizes=\"(max-width: 200px) 100vw, 200px\" /><img loading=\"lazy\"\n          sizes=\"(max-width: 200px) 100vw, 200px\"\n          srcset=\"https://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=50&h=75&q=50 50w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=100&h=150&q=50 100w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=200&h=300&q=50 200w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=300&h=450&q=50 300w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=400&h=600&q=50 400w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=600&h=900&q=50 600w\"\n          src=\"https://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=200&q=50\"\n          alt=\"\"\n          style=\"position:absolute;top:0;left:0;opacity:1;width:100%;height:100%;object-fit:cover;object-position:center\" />\n      </picture>\n    </noscript>\n  </div>\n  <p></p>\n  <hr>\n  <p>The languages only differ in:</p>\n  <ol>\n    <li>\n      <p>their grammar</p>\n    </li>\n    <li>\n      <p>their pronunciation</p>\n    </li>\n    <li>\n      <p>their most common words</p>\n    </li>\n    <li>\n      <p><span data-cy-id=\"inline-text\">[Inline-ContentfulText]\n          <!-- -->Text: Short\n          <!-- -->:\n          <!-- -->The quick brown fox jumps over the lazy dog.</span></p>\n    </li>\n  </ol>\n  <p>Everyone realizes why a new common language would be desirable: one could\n    refuse to pay expensive translators.</p>\n  <p><code>{\n      \"userId\": 1,\n      \"id\": 1,\n      \"title\": \"delectus aut autem\",\n      \"completed\": false\n      }</code></p>\n  <p>To achieve this, it would be necessary to have uniform grammar,\n    pronunciation and more common words.</p>\n  <p data-cy-id=\"location\">[ContentfulLocation] Lat:\n    <!-- -->52.51627\n    <!-- -->, Long:\n    <!-- -->13.3777</p>\n  <blockquote>\n    <p>If several languages coalesce, the grammar of the resulting language is\n      more simple and regular than that of the individual languages.</p>\n  </blockquote>\n  <p>The new common language will be more simple and regular than the existing\n    European languages. It will be as simple as Occidental; in fact, it will be\n  </p>\n  <hr>\n</div>"
+      "1": "<div data-cy-id=\"rich-text-all-features\">\n  <h2>Rich Text: All Features</h2>\n  <h1>The <a href=\"https://contentful.com/\">European</a> languages</h1>\n  <p>are members of the same family. Their <u>separate</u> <strong\n      data-cy-strong=\"true\">existence</strong> is a <i>myth</i>. For:</p>\n  <ul>\n    <li>\n      <p>science</p>\n    </li>\n    <li>\n      <p>music</p>\n    </li>\n    <li>\n      <p>sport</p>\n    </li>\n    <li>\n      <p>etc</p>\n    </li>\n  </ul>\n  <p>Europe uses the same vocabulary.</p>\n  <div data-gatsby-image-wrapper=\"\"\n    class=\"gatsby-image-wrapper gatsby-image-wrapper-constrained\">\n    <div><img alt=\"\"\n        role=\"presentation\"\n        aria-hidden=\"true\"\n        src=\"data:image/svg+xml;charset=utf-8,%3Csvg height='300' width='200' xmlns='http://www.w3.org/2000/svg' version='1.1'%3E%3C/svg%3E\">\n    </div>\n    <div aria-hidden=\"true\"\n      data:image=\"\"\n      redacted;base64,redacted=\"\"></div>\n    <picture>\n      <source type=\"image/webp\"\n        sizes=\"(min-width: 200px) 200px, 100vw\"\n        srcset=\"https://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=50&amp;h=75&amp;q=50&amp;fm=webp 50w,https://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=100&amp;h=150&amp;q=50&amp;fm=webp 100w,https://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=200&amp;h=300&amp;q=50&amp;fm=webp 200w,https://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=400&amp;h=600&amp;q=50&amp;fm=webp 400w\">\n      <img data-gatsby-image-ssr=\"\"\n        data-main-image=\"\"\n        sizes=\"(min-width: 200px) 200px, 100vw\"\n        decoding=\"async\"\n        loading=\"lazy\"\n        alt=\"\"\n        src=\"https://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=200&amp;h=300&amp;fl=progressive&amp;q=50&amp;fm=jpg\"\n        srcset=\"https://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=50&amp;h=75&amp;fl=progressive&amp;q=50&amp;fm=jpg 50w,https://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=100&amp;h=150&amp;fl=progressive&amp;q=50&amp;fm=jpg 100w,https://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=200&amp;h=300&amp;fl=progressive&amp;q=50&amp;fm=jpg 200w,https://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=400&amp;h=600&amp;fl=progressive&amp;q=50&amp;fm=jpg 400w\">\n    </picture><noscript>\n      <picture>\n        <source type=\"image/webp\"\n          srcSet=\"https://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=50&amp;h=75&amp;q=50&amp;fm=webp 50w,https://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=100&amp;h=150&amp;q=50&amp;fm=webp 100w,https://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=200&amp;h=300&amp;q=50&amp;fm=webp 200w,https://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=400&amp;h=600&amp;q=50&amp;fm=webp 400w\"\n          sizes=\"(min-width: 200px) 200px, 100vw\" /><img\n          data-gatsby-image-ssr=\"\"\n          data-main-image=\"\"\n          sizes=\"(min-width: 200px) 200px, 100vw\"\n          decoding=\"async\"\n          loading=\"lazy\"\n          src=\"https://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=200&amp;h=300&amp;fl=progressive&amp;q=50&amp;fm=jpg\"\n          srcSet=\"https://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=50&amp;h=75&amp;fl=progressive&amp;q=50&amp;fm=jpg 50w,https://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=100&amp;h=150&amp;fl=progressive&amp;q=50&amp;fm=jpg 100w,https://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=200&amp;h=300&amp;fl=progressive&amp;q=50&amp;fm=jpg 200w,https://images.ctfassets.net/k8iqpp6u0ior/2PeXF5HW2hiDfCggU9RIpO/8236da1f40af17e19922ddb87030735a/jae-park-7GX5aICb5i4-unsplash.jpg?w=400&amp;h=600&amp;fl=progressive&amp;q=50&amp;fm=jpg 400w\"\n          alt=\"\" /></picture>\n    </noscript>\n    <script type=\"module\">\n      const t=\"undefined\"!=typeof HTMLImageElement&&\"loading\"in HTMLImageElement.prototype;if(t){const t=document.querySelectorAll(\"img[data-main-image]\");for(let e of t){e.dataset.src&&(e.setAttribute(\"src\",e.dataset.src),e.removeAttribute(\"data-src\")),e.dataset.srcset&&(e.setAttribute(\"srcset\",e.dataset.srcset),e.removeAttribute(\"data-srcset\"));const t=e.parentNode.querySelectorAll(\"source[data-srcset]\");for(let e of t)e.setAttribute(\"srcset\",e.dataset.srcset),e.removeAttribute(\"data-srcset\");e.complete&&(e.style.opacity=1)}}\n    </script>\n  </div>\n  <p></p>\n  <hr>\n  <p>The languages only differ in:</p>\n  <ol>\n    <li>\n      <p>their grammar</p>\n    </li>\n    <li>\n      <p>their pronunciation</p>\n    </li>\n    <li>\n      <p>their most common words</p>\n    </li>\n    <li>\n      <p><span data-cy-id=\"inline-text\">[Inline-ContentfulText]\n          <!-- -->Text: Short\n          <!-- -->:\n          <!-- -->The quick brown fox jumps over the lazy dog.</span></p>\n    </li>\n  </ol>\n  <p>Everyone realizes why a new common language would be desirable: one could\n    refuse to pay expensive translators.</p>\n  <p><code>{\n      \"userId\": 1,\n      \"id\": 1,\n      \"title\": \"delectus aut autem\",\n      \"completed\": false\n      }</code></p>\n  <p>To achieve this, it would be necessary to have uniform grammar,\n    pronunciation and more common words.</p>\n  <p data-cy-id=\"location\">[ContentfulLocation] Lat:\n    <!-- -->52.51627\n    <!-- -->, Long:\n    <!-- -->13.3777</p>\n  <blockquote>\n    <p>If several languages coalesce, the grammar of the resulting language is\n      more simple and regular than that of the individual languages.</p>\n  </blockquote>\n  <p>The new common language will be more simple and regular than the existing\n    European languages. It will be as simple as Occidental; in fact, it will be\n  </p>\n  <hr>\n</div>"
     },
     "rich-text: Basic": {
       "1": "<div data-cy-id=\"rich-text-basic\">\n  <h2>Rich Text: Basic</h2>\n  <h1>The <a href=\"https://contentful.com/\">European</a> languages</h1>\n  <p>are members of the same family. Their <u>separate</u> <strong\n      data-cy-strong=\"true\">existence</strong> is a <i>myth</i>. For:</p>\n  <ul>\n    <li>\n      <p>science</p>\n    </li>\n    <li>\n      <p>music</p>\n    </li>\n    <li>\n      <p>sport</p>\n    </li>\n    <li>\n      <p>etc</p>\n    </li>\n  </ul>\n  <p>Europe uses the same vocabulary.</p>\n  <hr>\n  <p>The languages only differ in:</p>\n  <ol>\n    <li>\n      <p>their grammar</p>\n    </li>\n    <li>\n      <p>their pronunciation</p>\n    </li>\n    <li>\n      <p>their most common words</p>\n    </li>\n  </ol>\n  <p>Everyone realizes why a new common language would be desirable: one could\n    refuse to pay expensive translators.</p>\n  <p><code>{\n      \"userId\": 1,\n      \"id\": 1,\n      \"title\": \"delectus aut autem\",\n      \"completed\": false\n      }</code></p>\n  <p>To achieve this, it would be necessary to have uniform grammar,\n    pronunciation and more common words.</p>\n  <blockquote>\n    <p>If several languages coalesce, the grammar of the resulting language is\n      more simple and regular than that of the individual languages.</p>\n  </blockquote>\n  <p>The new common language will be more simple and regular than the existing\n    European languages. It will be as simple as Occidental; in fact, it will be\n  </p>\n  <hr>\n</div>"
@@ -37,7 +37,7 @@ module.exports = {
       "1": "<div data-cy-id=\"rich-text-embedded-entry\">\n  <h2>Rich Text: Embedded Entry</h2>\n  <h1>Embedded Entry</h1>\n  <p data-cy-id=\"text\">[ContentfulText]\n    <!-- -->The quick brown fox jumps over the lazy dog.</p>\n  <p></p>\n  <p></p>\n  <hr>\n</div>"
     },
     "rich-text: Embedded Asset": {
-      "1": "<div data-cy-id=\"rich-text-embedded-asset\">\n  <h2>Rich Text: Embedded asset</h2>\n  <h1>Embedded Asset</h1>\n  <div class=\" gatsby-image-wrapper\"\n    style=\"position:relative;overflow:hidden;width:200px\">\n    <div aria-hidden=\"true\"\n      style=\"width:100%;padding-bottom:66.67279749862055%\"></div><img\n      aria-hidden=\"true\"\n      src=\"data:image/redacted;base64,redacted\"\n      alt=\"\"\n      style=\"position: absolute; top: 0px; left: 0px; width: 100%; height: 100%; object-fit: cover; object-position: center center; opacity: 0; transition-delay: 500ms;\">\n    <picture>\n      <source\n        srcset=\"https://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=50&amp;h=33&amp;q=50 50w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=100&amp;h=67&amp;q=50 100w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=200&amp;h=133&amp;q=50 200w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=300&amp;h=200&amp;q=50 300w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=400&amp;h=267&amp;q=50 400w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=600&amp;h=400&amp;q=50 600w\"\n        sizes=\"(max-width: 200px) 100vw, 200px\"><img\n        sizes=\"(max-width: 200px) 100vw, 200px\"\n        srcset=\"https://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=50&amp;h=33&amp;q=50 50w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=100&amp;h=67&amp;q=50 100w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=200&amp;h=133&amp;q=50 200w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=300&amp;h=200&amp;q=50 300w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=400&amp;h=267&amp;q=50 400w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=600&amp;h=400&amp;q=50 600w\"\n        src=\"https://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=200&amp;q=50\"\n        alt=\"\"\n        loading=\"lazy\"\n        style=\"position: absolute; top: 0px; left: 0px; width: 100%; height: 100%; object-fit: cover; object-position: center center; opacity: 1; transition: opacity 500ms ease 0s;\">\n    </picture><noscript>\n      <picture>\n        <source\n          srcset=\"https://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=50&h=33&q=50 50w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=100&h=67&q=50 100w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=200&h=133&q=50 200w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=300&h=200&q=50 300w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=400&h=267&q=50 400w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=600&h=400&q=50 600w\"\n          sizes=\"(max-width: 200px) 100vw, 200px\" /><img loading=\"lazy\"\n          sizes=\"(max-width: 200px) 100vw, 200px\"\n          srcset=\"https://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=50&h=33&q=50 50w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=100&h=67&q=50 100w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=200&h=133&q=50 200w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=300&h=200&q=50 300w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=400&h=267&q=50 400w,\nhttps://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=600&h=400&q=50 600w\"\n          src=\"https://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=200&q=50\"\n          alt=\"\"\n          style=\"position:absolute;top:0;left:0;opacity:1;width:100%;height:100%;object-fit:cover;object-position:center\" />\n      </picture>\n    </noscript>\n  </div>\n  <p></p>\n  <p></p>\n  <p></p>\n  <hr>\n</div>"
+      "1": "<div data-cy-id=\"rich-text-embedded-asset\">\n  <h2>Rich Text: Embedded asset</h2>\n  <h1>Embedded Asset</h1>\n  <div data-gatsby-image-wrapper=\"\"\n    class=\"gatsby-image-wrapper gatsby-image-wrapper-constrained\">\n    <div><img alt=\"\"\n        role=\"presentation\"\n        aria-hidden=\"true\"\n        src=\"data:image/svg+xml;charset=utf-8,%3Csvg height='133' width='200' xmlns='http://www.w3.org/2000/svg' version='1.1'%3E%3C/svg%3E\">\n    </div>\n    <div aria-hidden=\"true\"\n      data:image=\"\"\n      redacted;base64,redacted=\"\"></div>\n    <picture>\n      <source type=\"image/webp\"\n        sizes=\"(min-width: 200px) 200px, 100vw\"\n        srcset=\"https://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=50&amp;h=33&amp;q=50&amp;fm=webp 50w,https://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=100&amp;h=67&amp;q=50&amp;fm=webp 100w,https://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=200&amp;h=133&amp;q=50&amp;fm=webp 200w,https://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=400&amp;h=266&amp;q=50&amp;fm=webp 400w\">\n      <img data-gatsby-image-ssr=\"\"\n        data-main-image=\"\"\n        sizes=\"(min-width: 200px) 200px, 100vw\"\n        decoding=\"async\"\n        loading=\"lazy\"\n        alt=\"\"\n        src=\"https://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=200&amp;h=133&amp;fl=progressive&amp;q=50&amp;fm=jpg\"\n        srcset=\"https://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=50&amp;h=33&amp;fl=progressive&amp;q=50&amp;fm=jpg 50w,https://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=100&amp;h=67&amp;fl=progressive&amp;q=50&amp;fm=jpg 100w,https://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=200&amp;h=133&amp;fl=progressive&amp;q=50&amp;fm=jpg 200w,https://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=400&amp;h=266&amp;fl=progressive&amp;q=50&amp;fm=jpg 400w\">\n    </picture><noscript>\n      <picture>\n        <source type=\"image/webp\"\n          srcSet=\"https://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=50&amp;h=33&amp;q=50&amp;fm=webp 50w,https://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=100&amp;h=67&amp;q=50&amp;fm=webp 100w,https://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=200&amp;h=133&amp;q=50&amp;fm=webp 200w,https://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=400&amp;h=266&amp;q=50&amp;fm=webp 400w\"\n          sizes=\"(min-width: 200px) 200px, 100vw\" /><img\n          data-gatsby-image-ssr=\"\"\n          data-main-image=\"\"\n          sizes=\"(min-width: 200px) 200px, 100vw\"\n          decoding=\"async\"\n          loading=\"lazy\"\n          src=\"https://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=200&amp;h=133&amp;fl=progressive&amp;q=50&amp;fm=jpg\"\n          srcSet=\"https://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=50&amp;h=33&amp;fl=progressive&amp;q=50&amp;fm=jpg 50w,https://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=100&amp;h=67&amp;fl=progressive&amp;q=50&amp;fm=jpg 100w,https://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=200&amp;h=133&amp;fl=progressive&amp;q=50&amp;fm=jpg 200w,https://images.ctfassets.net/k8iqpp6u0ior/3BSI9CgDdAn1JchXmY5IJi/f97a2185b3395591b98008647ad6fd3c/camylla-battani-AoqgGAqrLpU-unsplash.jpg?w=400&amp;h=266&amp;fl=progressive&amp;q=50&amp;fm=jpg 400w\"\n          alt=\"\" /></picture>\n    </noscript>\n    <script type=\"module\">\n      const t=\"undefined\"!=typeof HTMLImageElement&&\"loading\"in HTMLImageElement.prototype;if(t){const t=document.querySelectorAll(\"img[data-main-image]\");for(let e of t){e.dataset.src&&(e.setAttribute(\"src\",e.dataset.src),e.removeAttribute(\"data-src\")),e.dataset.srcset&&(e.setAttribute(\"srcset\",e.dataset.srcset),e.removeAttribute(\"data-srcset\"));const t=e.parentNode.querySelectorAll(\"source[data-srcset]\");for(let e of t)e.setAttribute(\"srcset\",e.dataset.srcset),e.removeAttribute(\"data-srcset\");e.complete&&(e.style.opacity=1)}}\n    </script>\n  </div>\n  <p></p>\n  <p></p>\n  <p></p>\n  <hr>\n</div>"
     },
     "rich-text: Embedded Entry With Deep Reference Loop": {
       "1": "<div data-cy-id=\"rich-text-embedded-entry-with-deep-reference-loop\">\n  <h2>Rich Text: Embedded entry with deep reference loop</h2>\n  <h1>Embedded entry with deep reference loop</h1>\n  <p data-cy-id=\"reference\">[ContentfulReference]\n    <!-- -->Content Reference: Many (2nd level loop)\n    <!-- -->: [\n    <!-- -->Number: Integer, Text: Short, Content Reference: One (Loop A -&gt;\n    B)\n    <!-- -->]</p>\n  <p></p>\n  <p></p>\n  <hr>\n</div>"
