+++ e2e-tests/contentful/src/pages/index.js
@@ -9,9 +9,6 @@ const IndexPage = () => (
     <h2>Media Rendering</h2>
     <ul>
       <li>
-        <Link to="/gatsby-image">/gatsby-image</Link>
-      </li>
-      <li>
         <Link to="/gatsby-plugin-image">/gatsby-plugin-image</Link>
       </li>
       <li>
